Move compute tags out of ErrorControl.hpp
Also rename IsUsingTimeSteppingErrorControl to
StepperErrorEstimatesEnabled as a more generic name.
