+++ src/Time/Tags/CMakeLists.txt
@@ -1,6 +1,12 @@
 # Distributed under the MIT License.
 # See LICENSE.txt for details.
 
+spectre_target_sources(
+  ${LIBRARY}
+  PRIVATE
+  StepperErrorTolerancesCompute.cpp
+  )
+
 spectre_target_headers(
   ${LIBRARY}
   INCLUDE_DIRECTORY ${CMAKE_SOURCE_DIR}/src
@@ -8,11 +14,12 @@ spectre_target_headers(
   AdaptiveSteppingDiagnostics.hpp
   FixedLtsRatio.hpp
   HistoryEvolvedVariables.hpp
-  IsUsingTimeSteppingErrorControl.hpp
   MinimumTimeStep.hpp
   StepChoosers.hpp
   StepNumberWithinSlab.hpp
+  StepperErrorEstimatesEnabled.hpp
   StepperErrorTolerances.hpp
+  StepperErrorTolerancesCompute.hpp
   StepperErrors.hpp
   Time.hpp
   TimeAndPrevious.hpp
