+++ tests/Unit/Time/Tags/Test_StepperErrorEstimatesEnabled.cpp
@@ -1,0 +1,15 @@
+// Distributed under the MIT License.
+// See LICENSE.txt for details.
+
+#include "Framework/TestingFramework.hpp"
+
+#include <string>
+
+#include "Helpers/DataStructures/DataBox/TestHelpers.hpp"
+#include "Time/Tags/StepperErrorEstimatesEnabled.hpp"
+
+SPECTRE_TEST_CASE("Unit.Time.Tags.StepperErrorEstimatesEnabled",
+                  "[Unit][Time]") {
+  TestHelpers::db::test_simple_tag<Tags::StepperErrorEstimatesEnabled>(
+      "StepperErrorEstimatesEnabled");
+}
